Implement sendDocumentSupport.
sendDocument is implemented as a special wrapper that decide
what content type to use based on the query content. Such approach
is not servant-client friendly, so we ToMultipart function becomes
partial, but it may be ok, as other solutions would require manual
copy of the types and it will be much harder to follow.

Fixes #30.