Fix response parsing in UpdatingMessages
Response to editMessage* requests is either "true", indicating that the
edited message has been inline, or a message otherwise. This was
(intuitively) represented with Either Bool Message. However, Aeson
understands this as a literal sum type of Left _ | Right _, which is not
the intended semantics. Replace Either with a custom data type which
exhibits the proper parsing behaviour.